adk_file_search_agent/requirements.txt
google-generativeai
python-dotenv
agent-development-kit
pypdf
